+++ kscan/src/iosMain/kotlin/org/ncgroup/kscan/view/ScannerViewImpl.ios.kt
@@ -27,6 +27,7 @@ import platform.AVFoundation.torchMode
 internal actual fun ScannerViewImpl(
     codeTypes: List<BarcodeFormat>,
     modifier: Modifier,
+    @Suppress("UNUSED_PARAMETER") cameraId: String?,
     scannerController: ScannerController?,
     filter: (Barcode) -> Boolean,
     autoZoom: Boolean,